Move back `_submit_setup.py` from being V1 only to being generic
